Customisations: Support dynamically-loaded customisations
Add support for two types of customisations:
- Settings -> Integrations -> Customisations settings pages
- Settings -> Device List -> (Product ID) settings pages

In future, we also plan to support:
- Navigation bar additions
- Quick access pane additions
- Quick access pane card additions

This commit includes a customisation compiler script which
generates a .json file which contains all of the compiled
(with rcc) data associated with a customisation.

It also includes models of customisations, which are loaded
at runtime from the filesystem (for GX and Desktop builds)
  /tmp/venus-gui-v2/customisations/ (GX builds)
  ~appdir/customisations/ (Desktop builds)
or from a given MQTT settings path (for WASM builds):
  /Gui2/Customisations

It also includes the hooks in gui-v2 to allow displaying
customisations appropriately where required.

Contributes to issue #1071